Education RA/requirements.txt
accelerate==1.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
ascii_colors==0.5.2
async-timeout==4.0.3
attrs==25.1.0
beautifulsoup4==4.13.3
bert-score==0.3.13
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deepsearch-glm==1.0.0
dill==0.3.9
distro==1.9.0
docling==2.16.0
docling-core==2.19.1
docling-ibm-models==3.3.2
docling-parse==3.3.1
easyocr==1.7.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.17.0
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
isort==6.0.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.17
langchain-core==0.3.35
langchain-text-splitters==0.3.6
langsmith==0.3.8
latex2mathml==3.77.0
lazy_loader==0.4
lightrag-hku==1.1.6
lxml==5.3.1
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
nano-vectordb==0.0.4.3
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==2.2.3
ollama==0.4.7
openai==1.64.0
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pdf2image==1.17.0
pillow==10.4.0
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
pipmaster==0.4.0
platformdirs==4.3.6
propcache==0.2.1
psutil==7.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pypandoc==1.15
pyparsing==3.2.1
pypdfium2==4.30.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.1
pywin32==307
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rouge==1.0.1
rpds-py==0.22.3
Rtree==1.3.0
safetensors==0.5.2
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semchunk==2.2.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1
tqdm==4.67.1
transformers==4.49.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
XlsxWriter==3.2.2
yarl==1.18.3
zstandard==0.23.0
